delete dead code, rename our command to replace theirs
