+++ src/main/java/seedu/todo/logic/commands/DeleteCommand.java
@@ -1,50 +1,25 @@
 package seedu.todo.logic.commands;
 
-import seedu.todo.commons.core.Messages;
-import seedu.todo.commons.core.UnmodifiableObservableList;
-import seedu.todo.model.person.ReadOnlyPerson;
-import seedu.todo.model.person.UniquePersonList.PersonNotFoundException;
+import seedu.todo.commons.exceptions.IllegalValueException;
+import seedu.todo.logic.arguments.Argument;
+import seedu.todo.logic.arguments.IntArgument;
+import seedu.todo.logic.arguments.Parameter;
+import seedu.todo.model.task.ImmutableTask;
 
-/**
- * Deletes a person identified using it's last displayed index from the address book.
- */
-public class DeleteCommand extends Command {
-
-    public static final String COMMAND_WORD = "delete";
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Deletes the person identified by the index number used in the last person listing.\n"
-            + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
-
-    public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
-
-    public final int targetIndex;
-
-    public DeleteCommand(int targetIndex) {
-        this.targetIndex = targetIndex;
-    }
-
+public class DeleteCommand extends BaseCommand {
+    // TODO: Consider either creating a new subclass for indices 
+    // or create constraints framework for adding additional validation 
+    // to ensure this is always a position, non-zero number
+    private Argument<Integer> index = new IntArgument("index").required();
 
     @Override
-    public CommandResult execute() {
-
-        UnmodifiableObservableList<ReadOnlyPerson> lastShownList = model.getFilteredPersonList();
-
-        if (lastShownList.size() < targetIndex) {
-            indicateAttemptToExecuteIncorrectCommand();
-            return new CommandResult(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
-        }
-
-        ReadOnlyPerson personToDelete = lastShownList.get(targetIndex - 1);
-
-        try {
-            model.deletePerson(personToDelete);
-        } catch (PersonNotFoundException pnfe) {
-            assert false : "The target person cannot be missing";
-        }
-
-        return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
+    protected Parameter[] getArguments() {
+        return new Parameter[]{ index };
     }
 
+    @Override
+    public void execute() throws IllegalValueException {
+        ImmutableTask toDelete = this.getTaskAt(index.getValue());
+        this.model.delete(toDelete);
+    }
 }
